+++ src/components/tables/safetyAndSecurityLists.tsx
@@ -3,9 +3,10 @@ import { useCallback, useState } from "react";
 import Sort from "../filterAndSort/sort";
 import DeleteConfirmation from "../infoModal/delete-confirmation";
 import ModalTemplate from "../modal";
-import AddEditRoomOption from "../room-extra-options/add-edit-options";
+import AddEditExtraOption from "../room-extra-options/add-edit-options";
+import AddEdit from "../amenities/addEdit";
 
-export default function RoomOptionTable({
+export default function SafetyAndSecurityList({
   header,
   data,
   title,
@@ -13,7 +14,7 @@ export default function RoomOptionTable({
   header: string[];
   data: {
     id: number;
-    categoryName: string;
+    title: string;
     description: string;
   }[];
   title: string;
@@ -21,7 +22,7 @@ export default function RoomOptionTable({
   const [currentPage, setCurrentPage] = useState(1);
 
   const [openDelete, setOpenDelete] = useState(false);
-  const [editRoomOption, setEditRoomOption] = useState(false);
+  const [editSafetyAndSecurity, setEditSafetyAndSecurity] = useState(false);
   const [isDeleting, setIsDeleting] = useState(false);
 
   const handleDelete = useCallback(async () => {
@@ -39,7 +40,7 @@ export default function RoomOptionTable({
       <div className="w-full rounded-lg border p-5 flex flex-col gap-5 overflow-x-auto">
         <div className=" w-full justify-between gap-6 flex items-center flex-col lg:flex-row">
           <h2 className="text-xl font-semibold">{title}</h2>
-          <Sort id="room-options" label="Sort Category" />{" "}
+          <Sort id="extra-options" label="Sort List" />{" "}
         </div>
         <table className=" w-full">
           <thead className="">
@@ -53,24 +54,24 @@ export default function RoomOptionTable({
             {data.map((request, index) => {
               return (
                 <tr key={request?.id} className=" border-b">
-                  <td>{request?.categoryName}</td>
+                  <td>{request?.title}</td>
                   <td>{request?.description}</td>
                   <td className=" group relative">
                     <span className=" p-2 text-lg">...</span>
                     <span className="z-10 text-center group-hover:flex hidden w-52 bg-white text-sm absolute right-0 top-0 rounded-lg shadow-lg flex-col">
                       <button
                         type="button"
-                        onClick={() => setEditRoomOption(true)}
+                        onClick={() => setEditSafetyAndSecurity(true)}
                         className=" p-3 px-4 hover:bg-primary/10 transition-all rounded-lg"
                       >
-                        Edit Room Option
+                        Edit
                       </button>
                       <button
                         type="button"
                         onClick={() => setOpenDelete(true)}
                         className="p-3 px-4 hover:bg-primary/10 transition-all rounded-lg"
                       >
-                        Delete Room Option
+                        Delete
                       </button>
                     </span>
                   </td>
@@ -90,7 +91,7 @@ export default function RoomOptionTable({
           }}
           setCurrentPage={setCurrentPage}
           isLoading={false}
-          label="Room Options"
+          label="Safety & Security"
         />
       </div>
       <DeleteConfirmation
@@ -98,18 +99,18 @@ export default function RoomOptionTable({
         setOpen={setOpenDelete}
         isLoading={isDeleting}
         confirmationHandler={handleDelete}
-        title="Delete Room Option"
-        description="Are you sure you want to delete this room option?"
+        title="Delete Safety and Security"
+        description="Are you sure you want to delete this safety and security?"
         btnTitle="Yes, I want to"
       />
       <ModalTemplate
-        open={editRoomOption}
-        setOpen={setEditRoomOption}
+        open={editSafetyAndSecurity}
+        setOpen={setEditSafetyAndSecurity}
         showXicon={true}
-        title="Edit Room Option"
+        title="Edit"
         className=" max-w-md"
       >
-        <AddEditRoomOption setOpenOption={setEditRoomOption} id="1" />
+        <AddEdit setOpen={setEditSafetyAndSecurity} id="1" />
       </ModalTemplate>
     </>
   );